task-4/pyproject.toml
[project]
name = "task4-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyarrow>=22.0.0",
    "pypdf2>=3.0.1",
    "streamlit",
    "openai-agents",
    "python-dotenv",
]